Move client/cli/util_exec.go to proc/exec.go
