+++ pkg/client/cli/cmds_intercept.go
@@ -27,6 +27,7 @@ import (
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cache"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/cliutil"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/extensions"
+	"github.com/telepresenceio/telepresence/v2/pkg/proc"
 )
 
 type interceptArgs struct {
@@ -262,9 +263,9 @@ func intercept(cmd *cobra.Command, args interceptArgs) error {
 				if args.dockerRun {
 					return is.runInDocker(ctx, is.cmd, args.cmdline)
 				}
-				return start(ctx, args.cmdline[0], args.cmdline[1:], true,
+				return proc.Start(ctx, args.cmdline[0], args.cmdline[1:], true,
 					cmd.InOrStdin(), cmd.OutOrStdout(), cmd.ErrOrStderr(),
-					envPairs(is.env)...)
+					proc.EnvPairs(is.env)...)
 			})
 		})
 	})
@@ -681,7 +682,7 @@ func (is *interceptState) runInDocker(ctx context.Context, cmd safeCobraCommand,
 	if dockerMount != "" {
 		ourArgs = append(ourArgs, "-v", fmt.Sprintf("%s:%s", is.mountPoint, dockerMount))
 	}
-	return start(ctx, "docker", append(ourArgs, args...), true, cmd.InOrStdin(), cmd.OutOrStdout(), cmd.ErrOrStderr())
+	return proc.Start(ctx, "docker", append(ourArgs, args...), true, cmd.InOrStdin(), cmd.OutOrStdout(), cmd.ErrOrStderr())
 }
 
 func (is *interceptState) writeEnvFile() error {
